Simplify SHOW CURRENT NAMESPACE command
