+++ sql/catalyst/src/main/scala/org/apache/spark/sql/catalyst/plans/logical/v2Commands.scala
@@ -637,15 +637,6 @@ case class RefreshTable(child: LogicalPlan) extends UnaryCommand {
 }
 
 /**
- * The logical plan of the SHOW CURRENT NAMESPACE command.
- */
-case class ShowCurrentNamespace() extends LeafCommand {
-  override val output: Seq[Attribute] = Seq(
-    AttributeReference("catalog", StringType, nullable = false)(),
-    AttributeReference("namespace", StringType, nullable = false)())
-}
-
-/**
  * The logical plan of the SHOW TBLPROPERTIES command.
  */
 case class ShowTableProperties(
